Move any & intersection from SymbolVisitors to Symbols
Follow up to https://github.com/crate/crate/pull/16236
