+++ server/src/main/java/io/crate/execution/engine/collect/DocValuesGroupByOptimizedIterator.java
@@ -92,7 +92,7 @@ final class DocValuesGroupByOptimizedIterator {
                                           DocInputFactory docInputFactory,
                                           RoutedCollectPhase collectPhase,
                                           CollectTask collectTask) {
-        if (Symbols.containsColumn(collectPhase.toCollect(), DocSysColumns.SCORE)
+        if (Symbols.hasColumn(collectPhase.toCollect(), DocSysColumns.SCORE)
             || collectPhase.where().hasColumn(DocSysColumns.SCORE)) {
             return null;
         }
