+++ server/src/main/java/io/crate/execution/engine/collect/GroupByOptimizedIterator.java
@@ -132,7 +132,7 @@ final class GroupByOptimizedIterator {
         if (!keyRef.hasDocValues()) {
             return null;
         }
-        if (Symbols.containsColumn(collectPhase.toCollect(), DocSysColumns.SCORE)
+        if (Symbols.hasColumn(collectPhase.toCollect(), DocSysColumns.SCORE)
             || collectPhase.where().hasColumn(DocSysColumns.SCORE)) {
             // We could optimize this, but since it's assumed to be an uncommon case we fallback to generic group-by
             // to keep the optimized implementation a bit simpler
