+++ server/src/main/java/io/crate/planner/operators/NestedLoopJoin.java
@@ -49,7 +49,6 @@ import io.crate.execution.dsl.projection.builder.InputColumns;
 import io.crate.execution.dsl.projection.builder.ProjectionBuilder;
 import io.crate.expression.symbol.SelectSymbol;
 import io.crate.expression.symbol.Symbol;
-import io.crate.expression.symbol.SymbolVisitors;
 import io.crate.expression.symbol.Symbols;
 import io.crate.planner.DependencyCarrier;
 import io.crate.planner.ExecutionPlan;
@@ -225,22 +224,22 @@ public class NestedLoopJoin extends AbstractJoinPlan {
         LinkedHashSet<Symbol> lhsToKeep = new LinkedHashSet<>();
         LinkedHashSet<Symbol> rhsToKeep = new LinkedHashSet<>();
         for (Symbol outputToKeep : outputsToKeep) {
-            SymbolVisitors.intersection(outputToKeep, lhs.outputs(), lhsToKeep::add);
-            SymbolVisitors.intersection(outputToKeep, rhs.outputs(), rhsToKeep::add);
+            Symbols.intersection(outputToKeep, lhs.outputs(), lhsToKeep::add);
+            Symbols.intersection(outputToKeep, rhs.outputs(), rhsToKeep::add);
         }
 
         if (joinCondition != null) {
             // If there a lookup-join in place, and the outputs belong only to the lookup side,
             // we can drop the join and return only the lookup-side
             if (lhsToKeep.isEmpty() && lookupJoin == LookUpJoin.RIGHT) {
-                SymbolVisitors.intersection(joinCondition, rhs.outputs(), rhsToKeep::add);
+                Symbols.intersection(joinCondition, rhs.outputs(), rhsToKeep::add);
                 return rhs.pruneOutputsExcept(rhsToKeep);
             } else if (rhsToKeep.isEmpty() && lookupJoin == LookUpJoin.LEFT) {
-                SymbolVisitors.intersection(joinCondition, lhs.outputs(), lhsToKeep::add);
+                Symbols.intersection(joinCondition, lhs.outputs(), lhsToKeep::add);
                 return lhs.pruneOutputsExcept(lhsToKeep);
             } else {
-                SymbolVisitors.intersection(joinCondition, lhs.outputs(), lhsToKeep::add);
-                SymbolVisitors.intersection(joinCondition, rhs.outputs(), rhsToKeep::add);
+                Symbols.intersection(joinCondition, lhs.outputs(), lhsToKeep::add);
+                Symbols.intersection(joinCondition, rhs.outputs(), rhsToKeep::add);
             }
         }
         LogicalPlan newLhs = lhs.pruneOutputsExcept(lhsToKeep);
@@ -266,12 +265,12 @@ public class NestedLoopJoin extends AbstractJoinPlan {
         LinkedHashSet<Symbol> usedFromLeft = new LinkedHashSet<>();
         LinkedHashSet<Symbol> usedFromRight = new LinkedHashSet<>();
         for (Symbol usedColumn : usedColumns) {
-            SymbolVisitors.intersection(usedColumn, lhs.outputs(), usedFromLeft::add);
-            SymbolVisitors.intersection(usedColumn, rhs.outputs(), usedFromRight::add);
+            Symbols.intersection(usedColumn, lhs.outputs(), usedFromLeft::add);
+            Symbols.intersection(usedColumn, rhs.outputs(), usedFromRight::add);
         }
         if (joinCondition != null) {
-            SymbolVisitors.intersection(joinCondition, lhs.outputs(), usedFromLeft::add);
-            SymbolVisitors.intersection(joinCondition, rhs.outputs(), usedFromRight::add);
+            Symbols.intersection(joinCondition, lhs.outputs(), usedFromLeft::add);
+            Symbols.intersection(joinCondition, rhs.outputs(), usedFromRight::add);
         }
         FetchRewrite lhsFetchRewrite = lhs.rewriteToFetch(usedFromLeft);
         FetchRewrite rhsFetchRewrite = rhs.rewriteToFetch(usedFromRight);
