+++ server/src/main/java/io/crate/planner/operators/ProjectSet.java
@@ -37,7 +37,7 @@ import io.crate.execution.dsl.projection.builder.InputColumns;
 import io.crate.execution.dsl.projection.builder.ProjectionBuilder;
 import io.crate.expression.symbol.Function;
 import io.crate.expression.symbol.Symbol;
-import io.crate.expression.symbol.SymbolVisitors;
+import io.crate.expression.symbol.Symbols;
 import io.crate.metadata.FunctionType;
 import io.crate.planner.DependencyCarrier;
 import io.crate.planner.ExecutionPlan;
@@ -151,10 +151,10 @@ public class ProjectSet extends ForwardingLogicalPlan {
         LinkedHashSet<Symbol> toKeep = new LinkedHashSet<>();
         LinkedHashSet<Symbol> newStandalone = new LinkedHashSet<>();
         for (Symbol outputToKeep : outputsToKeep) {
-            SymbolVisitors.intersection(outputToKeep, standalone, newStandalone::add);
+            Symbols.intersection(outputToKeep, standalone, newStandalone::add);
         }
         for (Function tableFunction : tableFunctions) {
-            SymbolVisitors.intersection(tableFunction, source.outputs(), toKeep::add);
+            Symbols.intersection(tableFunction, source.outputs(), toKeep::add);
         }
         toKeep.addAll(newStandalone);
         LogicalPlan newSource = source.pruneOutputsExcept(toKeep);
